+++ src/agendaTelefonica/Contacto/Contacto.java
@@ -4,13 +4,13 @@ public class Contacto {
 
     protected String nombre;
     protected String apellido;
-    protected int telefono;
+    protected long telefono;
 
     public Contacto(){
 
     }
 
-    public Contacto(String nombre, String apellido, int telefono) {
+    public Contacto(String nombre, String apellido, long telefono) {
         this.nombre = nombre;
         this.apellido = apellido;
         this.telefono = telefono;
@@ -33,11 +33,11 @@ public class Contacto {
         this.apellido = apellido;
     }
 
-    public int getTelefono() {
+    public long getTelefono() {
         return telefono;
     }
 
-    public void setTelefono(int telefono) {
+    public void setTelefono(long telefono) {
         this.telefono = telefono;
     }
 }
